Add possibility to add a plugin signature
This makes the documentation of the plugins much better.
It shouldn't break existing plugins besides maybe waste performance
while throwing errors.
The only bad side effect is that maybe if a plugin is triggered just by
calling it there will be a bad call and that would mean a signature check
needs to be added.