Introduce new TwitchChatHandler type that heavily reduces LOC
Because now all the types can just be imported via the type, all changes
directly influence all uses of it.

Also all file names are now defined in an external file for less breaks in
the future if changes are made.

The type guards for the 2 custom JSON files do now report to the console
in more detail about what is bad about the failed to loaded files.