Add dynamic database storage for persistent map request demands
This should remove a lot of dynamic information from the static .env file and not
only save it persistently but also make it easy to change at runtime without being
reset on a restart.
In turn of these changes the whole database file layout was updated to allow
multiple databases to be used much more intuitively and easily.