Automate GitHub releases and decouple documentation generation from CI
This means no longer do files need to be downloaded, renamed and reuploaded
in what takes more than an hour and in which someone could modify the binaries
but now everything will be renamed and uploaded by GitHub Actions itself.
Only the title and description need to be edited if need be.
Also the documentation can be triggered manually at any time and is no longer
using a 3rd party addon or a branch but uses the new GitHub pages API.